+++ src/app/app.routes.ts
@@ -4,7 +4,7 @@ import { WelcomePage } from "./pages/welcome/welcome"
 import { ProsodyPage } from "./pages/new-poem/step-1/ProsodyPage"
 import { MivesPage } from "./pages/new-poem/step-2/MivesPage"
 import { WeightsPage } from "./pages/new-poem/step-3/WeightsPage"
-import { PoemPage } from "./pages/poem/poem.component"
+import { PoemPage } from "./pages/poem/PoemPage"
 
 export const routes: Routes = [
     { path: "", component: WelcomePage },
